--- students/0000009/0000009-p01/src/main/java/ru/mirea/practice/s0000009/ArgClass.java
@@ -1,12 +1,0 @@
-package ru.mirea.practice.s0000009;
-
-//Третья задача 1 практической работы
-public class ArgClass
-{
-    public static void main(String[] args)
-    {
-        for (String str : args){
-            System.out.println("аргумент = " + str);
-        }
-    }
-}
