--- students/0000009/0000009-p02/src/main/java/ru/mirea/practice2/s0000009/TestBall.java
@@ -1,12 +1,0 @@
-package ru.mirea.practice2.s0000009;
-public class TestBall {
-    public static void main(String[] args)
-    {
-        Ball obj1 = new Ball();
-        Ball obj2 = new Ball(2.5,3.4);
-        System.out.println(obj1);
-        System.out.println(obj2);
-        obj1.setXY(32,33);
-        System.out.println(obj1);
-    }
-}
