--- students/0000009/0000009-p02/src/main/java/ru/mirea/practice5/s0000009/DogsHouse.java
@@ -1,23 +1,0 @@
-package ru.mirea.practice5.s0000009;
-
-import java.util.ArrayList;
-
-public class DogsHouse
-{
-    private ArrayList<Dog> dogs = new ArrayList<Dog>();
-    public void AddDog(Dog obj)
-    {
-        dogs.add(obj);
-    }
-
-    public void setDogs(ArrayList<Dog> dogs) {
-        this.dogs = dogs;
-    }
-
-    public ArrayList<Dog> getDogs() {
-        return dogs;
-    }
-    public String toString(){
-        return "Массив собак:" + dogs ;
-    }
-}
