--- students/0000009/0000009-p02/src/main/java/ru/mirea/practice6/s0000009/CircleTest.java
@@ -1,13 +1,0 @@
-package ru.mirea.practice6.s0000009;
-
-public class CircleTest
-{
-    public static void main(String[] args) {
-        Circle obj1 = new Circle(1,1,10,"green");
-        Circle obj2 = new Circle(2,3,4,"blue");
-        System.out.println(obj1.AreaCircle());
-        System.out.println(obj1.LengthCircle());
-        System.out.println(obj1);
-        System.out.println(obj1.MaxCircle(obj1,obj2));
-    }
-}
