--- students/0000009/pom.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <artifactId>algorithms-and-data-structures</artifactId>
-        <groupId>ru.mirea.practice</groupId>
-        <version>2022.1</version>
-        <relativePath>../../pom.xml</relativePath>
-    </parent>
-    <artifactId>0000009</artifactId>
-    <packaging>pom</packaging>
-    <description>Практическая работа Турчанинова Даниила</description>
-
-    <modules>
-        <module>0000009-p01</module>
-        <module>0000009-p02</module>
-    </modules>
-</project>
