+++ students/21K0709/21K0709-p01/src/main/java/ru/mirea/practice/s21K0709/Array.java
@@ -1,0 +1,21 @@
+package ru.mirea.practice.s21K0709;
+
+//Первая задача 1 практической работы
+public class Array
+{
+    public static void main(String[] args) {
+        int [] mas = new int[10];
+        int sum = 0;
+        for (int i = 0;i<mas.length;i++)
+        {
+            mas[i] = i;
+            System.out.print(i);
+            System.out.print(" ");
+            sum+=i;
+        }
+        System.out.println();
+        System.out.println(sum/ mas.length);
+    }
+
+}
+
