+++ students/21K0709/21K0709-p02/src/main/java/ru/mirea/practice5/s21K0709/DogsHouse.java
@@ -1,0 +1,23 @@
+package ru.mirea.practice5.s21K0709;
+
+import java.util.ArrayList;
+
+public class DogsHouse
+{
+    private ArrayList<Dog> dogs = new ArrayList<Dog>();
+    public void AddDog(Dog obj)
+    {
+        dogs.add(obj);
+    }
+
+    public void setDogs(ArrayList<Dog> dogs) {
+        this.dogs = dogs;
+    }
+
+    public ArrayList<Dog> getDogs() {
+        return dogs;
+    }
+    public String toString(){
+        return "Массив собак:" + dogs ;
+    }
+}
