+++ students/21K0709/pom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <artifactId>algorithms-and-data-structures</artifactId>
+        <groupId>ru.mirea.practice</groupId>
+        <version>2022.1</version>
+        <relativePath>../../pom.xml</relativePath>
+    </parent>
+    <artifactId>21K0709</artifactId>
+    <packaging>pom</packaging>
+    <description>Практическая работа Турчанинова Даниила</description>
+
+    <modules>
+        <module>21K0709-p01</module>
+        <module>21K0709-p02</module>
+    </modules>
+</project>
